ISSUE #3776 - move all selectors together
